Utility program now adds correct permissions & kit pages

+++ Source/Domains/C8S.Domain.Features/OrderOffers/Models/OrderOffer.cs
@@ -6,7 +6,7 @@ public record OrderOffer
 {
     public int OrderOfferId { get; init; }
     public int Ordinal { get; init; }
-    public short Quantity { get; init; }
+    public int Quantity { get; init; }
 
     public Offer Offer { get; init; } = null!;
 }
